requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
librosa>=0.10.0
scikit-image>=0.21.0
lime>=0.2.0.1
shap>=0.42.0
